feat(nexus): migrate graph endpoint to hexagonal architecture
Closes #846.

Extracts all business logic from the monolithic graph.py endpoint into a
proper hexagonal service following the same ports-and-adapters pattern
established by chat, search, agents, and other services.

New structure under services/graph/:
- graph__schema.py — frozen domain dataclasses + domain exceptions
  (GraphServiceUnavailableError, GraphProtectedError)
- service.py — GraphService with triple-store-backed operations
  (list, create, clear, delete, overview, network)
- adapters/primary/ — FastAPI router, Pydantic HTTP schemas, DI helper
- handlers/ — re-exports router (convention)

Updated:
- registry.py: added graph: GraphService field + property
- registry_bootstrap.py: wires GraphService() into the registry
- exceptions.py: registers new domain errors in the global handler
- api/endpoints/graph.py: replaced with a compatibility shim
  (mirrors the search.py pattern; existing importers like view.py
  continue to resolve GraphEdge/GraphNode unchanged)
