+++ libs/naas-abi/naas_abi/apps/nexus/apps/api/app/services/graph/handlers/__init__.py
@@ -1,0 +1,5 @@
+from naas_abi.apps.nexus.apps.api.app.services.graph.handlers.graph__http_handler import (
+    router,
+)
+
+__all__ = ["router"]
