+++ libs/naas-abi/naas_abi/apps/nexus/apps/api/app/services/graph/handlers/graph__http_handler.py
@@ -1,0 +1,3 @@
+from naas_abi.apps.nexus.apps.api.app.services.graph.adapters.primary import router
+
+__all__ = ["router"]
