+++ libs/naas-abi/naas_abi/apps/nexus/apps/api/app/services/registry_bootstrap.py
@@ -11,6 +11,7 @@ from naas_abi.apps.nexus.apps.api.app.services.chat.adapters.secondary.postgres 
     ChatSecondaryAdapterPostgres,
 )
 from naas_abi.apps.nexus.apps.api.app.services.chat.service import ChatService
+from naas_abi.apps.nexus.apps.api.app.services.graph.service import GraphService
 from naas_abi.apps.nexus.apps.api.app.services.iam.adapters.secondary.postgres import (
     IAMSecondaryAdapterPostgres,
 )
@@ -53,6 +54,7 @@ def initialize_nexus_service_registry() -> ServiceRegistry:
         AgentSecondaryAdapterPostgres(db_getter=db_getter),
         iam_service=iam_service,
     )
+    graph_service = GraphService()
 
     return ServiceRegistry.configure(
         RegistryServices(
@@ -62,5 +64,6 @@ def initialize_nexus_service_registry() -> ServiceRegistry:
             agents=agents_service,
             workspaces=workspace_service,
             organizations=organization_service,
+            graph=graph_service,
         )
     )
